implement parsing of all motion blocks (though generation is not done yet), parse options directly instead of going through intermediary blocks (like CloneMenu, GotoMenu, etc...), fix expectation of integer coordinates and direction

+++ runtime.h
@@ -23,10 +23,10 @@ typedef struct {
 } Sprite;
 
 typedef struct {
-    int x;
-    int y;
-    int size;
-    int direction;
+    float x;
+    float y;
+    float size;
+    float direction;
 	bool visible;
     int sprite_index;
 	Sprite *sprites;
